\name{jSDM-package}
\alias{jSDM}
\docType{package}

\title{joint species distribution models}

\description{\code{jSDM} is an R package for fitting joint species distribution models in a          hierarchical Bayesian framework.}

\details{
  \tabular{ll}{
    Package: \tab jSDM\cr
    Type: \tab Package\cr
    Version: \tab 0.1.0\cr
    Date: \tab 2019-05-20\cr
    License: \tab GPL-3\cr
    LazyLoad: \tab yes\cr
  }
}

\author{
  \tabular{l}{
    Ghislain Vieilledent <ghislain.vieilledent@cirad.fr>\cr
    Jeanne Clément <jeanne.clement16@laposte.net>\cr
  }
}

\keyword{biodiversity}
\keyword{biotic interactions}
\keyword{community ecology}
\keyword{Gibbs sampling}
\keyword{hierarchical Bayesian models}
\keyword{MCMC}
\keyword{Markov Chains Monte Carlo}
\keyword{probit}
\keyword{species distribution models}

