\name{plot_residual_cor}

\alias{plot_residual_cor}
%- Also NEED an '\alias' for EACH other topic documented here.

\title{Plot the residual correlation matrix from a LVM.}

\description{
Plot the posterior mean estimator of residual correlation matrix reordered by first principal component.}

\usage{
plot_residual_cor(mod)
}

%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{mod}{
    An object of class \code{"jSDM"}.
  }
}

\author{
  Jeanne Clement \email{jeanne.clement16@laposte.net}
  Ghislain Vieilledent \email{ghislain.vieilledent@cirad.fr}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{corrplot}% __ONLY ONE__ keyword per line
