% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jSDM-package.R
\docType{package}
\name{jSDM-package}
\alias{jSDM-package}
\title{joint species distribution models}
\description{
\code{jSDM} is an R package for fitting joint species distribution models (jSDM) in a hierarchical Bayesian framework. The Gibbs sampler is written in C++. It uses Rcpp, Armadillo and GSL to maximize computation efficiency.
The package includes the following functions to fit various species distribution models :
}
\details{
\tabular{ll}{
   Package: \tab jSDM\cr
   Type: \tab Package\cr
   Version: \tab 0.1.0\cr
   Date: \tab 2019-01-11\cr
   License: \tab GPL-3 \cr
   LazyLoad: \tab yes\cr }
}
\author{
\tabular{l}{
Ghislain Vieilledent <ghislain.vieilledent@cirad.fr>\cr
Jeanne Clément <jeanne.clement16@laposte.net>\cr
}
}
\keyword{MCMC,}
\keyword{Metropolis,}
\keyword{algorithm,}
\keyword{binomial,}
\keyword{biodiversity,}
\keyword{logistic}
\keyword{model}
\keyword{multivariate}
\keyword{regression}
