\name{jSDM-package}
\alias{jSDM}
\docType{package}

\title{joint species distribution models}

\description{\code{jSDM} is an R package for fitting joint species distribution models in a          hierarchical Bayesian framework.}

\details{
  \tabular{ll}{
    Package: \tab jSDM\cr
    Type: \tab Package\cr
    Version: \tab 0.1.0\cr
    Date: \tab 2019-01-11\cr
    License: \tab GPL-3 \cr
    LazyLoad: \tab yes\cr
  }
}

\author{
  \tabular{l}{
    Ghislain Vieilledent <ghislain.vieilledent@cirad.fr>\cr
    Matthieu Authier <matthieu.authier@univ-lr.fr>\cr
  }
}

\keyword{biodiversity}
\keyword{species distribution models}
\keyword{hierarchical Bayesian models}
\keyword{spatial correlation}
\keyword{intrinsic CAR model}
\keyword{conditional autoregressive model}
\keyword{MCMC}
\keyword{Markov Chains Monte Carlo}
\keyword{Metropolis algorithm}

