\name{plot_residual_cor}

\alias{plot_residual_cor}

\title{Plot the residual correlation matrix from a LVM.}

\description{
Plot the posterior mean estimator of residual correlation matrix reordered by first principal component.}

\usage{
plot_residual_cor(mod)
}

\arguments{
  \item{mod}{
    An object of class \code{"jSDM"}.
  }
}

\author{
  Jeanne Clement \email{jeanne.clement16@laposte.net}
}
\keyword{corrplot}
